drone_controller.py
#!/usr/bin/env python3

import time
import threading
import logging
from typing import Dict, Optional
import RPi.GPIO as GPIO
from adxl345 import ADXL345
from vl53l0x import VL53L0X
import motor
import flight
import camera
import servos

class DroneState:
    IDLE = "IDLE"
    ARMED = "ARMED"
    FLYING = "FLYING"
    EMERGENCY = "EMERGENCY"
    LANDING = "LANDING"

class DroneController:
    def __init__(self):
        # Initialize logging
        logging.basicConfig(
            level=logging.INFO,
            format='%(asctime)s - %(levelname)s - %(message)s',
            handlers=[
                logging.FileHandler('drone.log'),
                logging.StreamHandler()
            ]
        )
        self.logger = logging.getLogger(__name__)
        
        # Drone state
        self.state = DroneState.IDLE
        self.armed = False
        self.battery_level = 100
        self.altitude = 0
        self.is_emergency = False
        
        # Initialize sensors
        try:
            self.accelerometer = ADXL345()
            self.distance_sensor = VL53L0X()
            self.distance_sensor.start_ranging()
        except Exception as e:
            self.logger.error(f"Failed to initialize sensors: {str(e)}")
            self.is_emergency = True
        
        # Initialize GPIO
        GPIO.setmode(GPIO.BOARD)
        self.setup_motors()
        
        # Start monitoring threads
        self.start_monitoring()
    
    def setup_motors(self):
        """Initialize all motors with safe starting values"""
        try:
            # Initialize your motors here
            # Example: self.motors = [motor.Motor(pin) for pin in motor_pins]
            self.logger.info("Motors initialized successfully")
        except Exception as e:
            self.logger.error(f"Motor initialization failed: {str(e)}")
            self.is_emergency = True
    
    def arm(self) -> bool:
        """Arm the drone if all systems are go"""
        if self.is_emergency:
            self.logger.error("Cannot arm: Drone is in emergency state")
            return False
            
        if self.battery_level < 20:
            self.logger.error("Cannot arm: Battery too low")
            return False
            
        self.armed = True
        self.state = DroneState.ARMED
        self.logger.info("Drone armed successfully")
        return True
    
    def disarm(self):
        """Safely disarm the drone"""
        self.armed = False
        self.state = DroneState.IDLE
        # Stop all motors
        self.logger.info("Drone disarmed")
    
    def emergency_stop(self):
        """Immediate stop of all motors"""
        self.is_emergency = True
        self.state = DroneState.EMERGENCY
        self.armed = False
        # Immediately stop all motors
        self.logger.warning("EMERGENCY STOP ACTIVATED")
    
    def start_monitoring(self):
        """Start background monitoring threads"""
        self.monitor_thread = threading.Thread(target=self._monitor_loop, daemon=True)
        self.monitor_thread.start()
    
    def _monitor_loop(self):
        """Background monitoring of drone systems"""
        while True:
            try:
                # Check battery
                if self.battery_level < 15:
                    self.logger.warning("Battery level critical")
                    self.initiate_landing()
                
                # Check accelerometer for unusual orientation
                accel_data = self.accelerometer.read()
                if abs(accel_data['x']) > 45 or abs(accel_data['y']) > 45:
                    self.logger.warning("Unusual orientation detected")
                    self.emergency_stop()
                
                # Monitor altitude
                altitude = self.distance_sensor.get_distance()
                if altitude > 100:  # Max height 100cm
                    self.logger.warning("Maximum altitude exceeded")
                    self.adjust_altitude(90)  # Move to safe altitude
                
                time.sleep(0.1)  # Check every 100ms
                
            except Exception as e:
                self.logger.error(f"Monitoring error: {str(e)}")
                self.emergency_stop()
    
    def adjust_altitude(self, target_altitude: float):
        """Adjust drone altitude to target (in cm)"""
        if not self.armed or self.is_emergency:
            return
            
        current_altitude = self.distance_sensor.get_distance()
        if current_altitude < target_altitude:
            flight.moveUp()
        elif current_altitude > target_altitude:
            flight.moveDown()
    
    def initiate_landing(self):
        """Begin automatic landing sequence"""
        self.state = DroneState.LANDING
        self.logger.info("Initiating landing sequence")
        
        while self.distance_sensor.get_distance() > 10:
            flight.moveDown()
            time.sleep(0.5)
        
        self.disarm()
    
    def get_telemetry(self) -> Dict:
        """Get current drone telemetry data"""
        return {
            'state': self.state,
            'armed': self.armed,
            'battery': self.battery_level,
            'altitude': self.distance_sensor.get_distance(),
            'accelerometer': self.accelerometer.read(),
            'emergency': self.is_emergency
        }

if __name__ == "__main__":
    drone = DroneController()
    # Add your control loop or interface here 